backend/requirements.txt
Flask
language-tool-python
psycopg2-binary
phonenumbers
pyjwt